create dataframe_exportable
